Moved render stuff, most of the game logic, and the game data from the view controller to the portable c framework.
A few issues remain
- IWFileTools: FileToString does not work correctly, something with the memory access I guess
- Update still remains in the ViewController for now

+++ IWGRenderer/IWUIRectangleButton.h
@@ -9,6 +9,8 @@
 #ifndef theBr1dge_IWUIRectangleButton_h
 #define theBr1dge_IWUIRectangleButton_h
 
+#include <stdbool.h>
+
 #include <OpenGLES/ES2/gl.h>
 #include <OpenGLES/ES2/glext.h>
 
@@ -46,6 +48,6 @@ IWUIRectangleButton IWUIRectangleButtonMake(float anchorPointX, float anchorPoin
                                           enum IWUIRECTANGLEBUTTON_CORNER_CUT cornerCut,
                                           float cornerOffset, float aspectRatio);
 
-int IWUIRectangleButtonToTriangleBuffer(IWUIRectangleButton *button, GLfloat* p);
+size_t IWUIRectangleButtonToTriangleBuffer(IWUIRectangleButton *button, GLfloat* p);
 
 #endif
